+++ genium/src/main/java/com/here/genium/model/cpp/CppField.kt
@@ -24,12 +24,13 @@ import java.util.stream.Stream
 
 class CppField(
     name: String,
+    fullyQualifiedName: String,
     type: CppTypeRef,
     val initializer: CppValue? = null,
     val isNotNull: Boolean = false,
     val isNullable: Boolean = false,
     val hasImmutableType: Boolean = false
-) : CppTypedElement(name, type) {
+) : CppTypedElement(name, fullyQualifiedName, type) {
 
     override fun stream() = Stream.of(type, initializer)